feat: created constants file feat: created ActivitiesSearchByDropDown component feat: moved RECORDS_PER_PAGE constant to constants file
