+++ src/renderer/components/blocks/widgets/index.ts
@@ -1,3 +1,4 @@
 export * from './ThemeModeSelector';
 export * from './SearchBox';
 export * from './ClimateActionModeRenderer';
+export * from './ActivitiesSearchByDropDown';
