+++ src/renderer/utils/constants.ts
@@ -1,0 +1,5 @@
+import { ActivitySearchBy } from '@/api';
+
+export const SEARCH_BY_CLIMATE_DATA: ActivitySearchBy = 'climate_warehouse';
+export const SEARCH_BY_ON_CHAIN_METADATA: ActivitySearchBy = 'onchain_metadata';
+export const RECORDS_PER_PAGE = 10;
